Revert "Refactor state components (#338)"
This reverts commit 0a59db56c87dc3c8cb89fc61450c45a0172427bd.
