--- .eslintrc.js
@@ -1,25 +1,0 @@
-module.exports = {
-  parser: 'babel-eslint',
-  env: {
-    browser: true,
-    es6: true,
-    node: true
-  },
-  extends: ['eslint:recommended', 'plugin:react/recommended'],
-  parserOptions: {
-    ecmaFeatures: {
-      experimentalObjectRestSpread: true,
-      jsx: true
-    },
-    sourceType: 'module'
-  },
-  plugins: ['react', 'import'],
-  rules: {
-    'react/react-in-jsx-scope': 'off',
-    'react/prop-types': 'off',
-    'react/display-name': 'off',
-    'react/jsx-uses-react': 'error',
-    'react/jsx-uses-vars': 'error',
-    'import/no-unresolved': 2
-  }
-}
