+++ lib/custom/modes/apache.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 // CodeMirror, copyright (c) by Marijn Haverbeke and others
 // Distributed under an MIT license: http://codemirror.net/LICENSE
 // Apache mode by gloony
