+++ lib/custom/modes/elixir.js
@@ -1,5 +1,4 @@
-/* eslint-disable */
 import CodeMirror from 'codemirror'
 
 // Require Codemirror elixir mode from npm modules and register it here
-export * from 'codemirror-mode-elixir'
+import registerElixirMode from 'codemirror-mode-elixir'
