+++ lib/custom/modes/graphql.js
@@ -1,3 +1,2 @@
-/* eslint-disable */
 import CodeMirror from 'codemirror'
 import 'codemirror-graphql/mode'
