+++ lib/custom/modes/nimrod.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 const CodeMirror = require('codemirror')
 
 CodeMirror.defineMode('nimrod', function(conf, parserConf) {
