+++ lib/customDomToImage.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 ;(function(global) {
   'use strict'
