+++ lib/util.js
@@ -36,7 +36,3 @@ export const fileToDataURL = blob =>
     reader.onload = e => res(e.target.result)
     reader.readAsDataURL(blob)
   })
-
-export function isImage(file) {
-  return file.type.split('/')[0] === 'image'
-}
