+++ package.json
@@ -10,7 +10,7 @@
     "test": "npm run cy:run --",
     "deploy": "now --public --docker -e NODE_ENV=production",
     "prettier": "prettier --config .prettierrc --write *.js {components,handlers,lib,pages}/*.js",
-    "lint": "eslint .",
+    "lint": "prettier --config .prettierrc *.js {components,handlers,lib,pages}/*.js",
     "precommit": "npm run contrib:build && git add README.md && lint-staged",
     "contrib:add": "all-contributors add",
     "contrib:build": "all-contributors generate",
@@ -23,11 +23,13 @@
     "codemirror": "^5.36.0",
     "codemirror-graphql": "^0.6.12",
     "codemirror-mode-elixir": "^1.1.1",
+    "cross-env": "^5.1.3",
     "dom-to-image": "^2.5.2",
     "downshift": "^1.28.0",
     "dropperx": "^0.1.0",
     "express": "^4.16.2",
-    "graphql": "^0.11.7",
+    "form-data": "^2.2.0",
+    "graphql": "^0.13.2",
     "highlight.js": "^9.12.0",
     "history": "^4.7.2",
     "isomorphic-fetch": "^2.2.1",
@@ -48,22 +50,18 @@
     "react-dnd-html5-backend": "^2.4.1",
     "react-dom": "16.2.0",
     "react-image-crop": "^3.0.9",
+    "react-spinkit": "^3.0.0",
     "react-spinner": "^0.2.7",
     "react-syntax-highlight": "^15.3.1",
     "resize-observer-polyfill": "^1.5.0",
     "tohash": "^1.0.2",
     "twitter": "^1.7.1",
-    "unsplash-js": "^4.8.0",
-    "unstated": "^2.0.2"
+    "unsplash-js": "^4.8.0"
   },
   "devDependencies": {
     "@zeit/next-css": "^0.1.5",
     "all-contributors-cli": "^4.7.0",
-    "babel-eslint": "8",
     "cypress": "^2.1.0",
-    "eslint": "^4.19.1",
-    "eslint-plugin-import": "^2.11.0",
-    "eslint-plugin-react": "^7.7.0",
     "hex2rgb": "^2.2.0",
     "husky": "^0.14.3",
     "lint-staged": "^7.0.4",
