+++ pages/_document.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import Document, { Head, Main, NextScript } from 'next/document'
 
 export default class extends Document {
