+++ pages/about.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import Page from '../components/Page'
 import { COLORS } from '../lib/constants'
