+++ server.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 const express = require('express')
 const morgan = require('morgan')
 const bodyParser = require('body-parser')
